Pangu/AImodels-raw/requirements_gpu.txt
numpy
onnx == 1.12.0
onnxruntime-gpu==1.19.2